requirements.txt
# Dependencies — Phase 2 (deterministic ILO rule engine + tracked changes).
# Each new phase adds its own libraries.

# Phase 1 — Streamlit UI + Word I/O
streamlit>=1.37.0
python-docx>=1.1.2
python-dotenv>=1.0.1

# Phase 2 — rule definitions in YAML
pyyaml>=6.0

# Phase 3 — LLM grammar/flow pass
google-generativeai>=0.8.0
groq>=0.11.0
requests>=2.31.0    # for Cerebras + Mistral REST calls

# Phase 5 — print-ready PDF
weasyprint>=62.0
jinja2>=3.1.4
